Add context menu
Add a context menu to editable fields in file pages (for testing
purposes for now) that appends the currently configured signature in
case the focused editable field is a textarea.

Apparently, adding the context menu entry only for textareas is not
possible - at least according to the documentation I found [1].
Therefore this commit adds a context menu entry for all editable fields
but drops requests made from non-textarea fields.

Fixes: #1

[1]: https://developer.mozilla.org/en-US/docs/Mozilla/Add-ons/WebExtensions/API/menus/ContextType