Adopt Photon Design by Firefox
Firefox has put out design guidelines that should be followed when
developing extensions and plugins [1]. These guidelines are adopted
for the plugin popup.

Also, the buttons for storing and resetting the signature text in the
popup seemed like overengineering, so I removed them. The signature will
now auto-update while the user edits the signature field in the popup.