Fix undefined signature on first use
A bug caused the extension popup to show the false signature value
'undefined' on first use of the plugin, rather than an empty field for
the signature. This is fixed by this change by checking the loaded value
received from the storage API before displaying it in the popup.

Also, nothing is appended in input fields if no signature is set (not
even linebreaks).