Add custom add-on ID for persistency testing
This change adds a custom add-on ID to the manifest file to persist
local storage for the extension. Without the explicit setting, the
extension will get a new ID assigned on every reload and any locally
stored data managed by the extension will be lost.

This explicit setting of the extension ID is only necessary for local
testing as the published extension will receive a dedicated add-on ID
once it is signed.