Exclude web-ext build directory from git
The previous commit falsely only added a single file of the web-ext
build directory to the .gitignore file. This is fixed here by adding
the full directory to the .gitignore file so no release artifacts are
tracked in the git repository. Tracking releases is done by tagging the
respective commits on the master branch.